Minor polish
"rigidbody" in the fan hover script was hiding the inherited rigidbody so I changed it to rb.


Added a priority variable to the rotate camera trigger script
Previously the active camera was being changed by enabling the next camera and all of them had the same priority so the active camera depended on how unity was rendering them and it had some randomness.