SoundManager improvements
Can now play multiple sounds of the same type at a time.

Plays the sounds in world space.

Creates gameobjects for playing the sounds, object pooling would be better but it's a marginal performance improvement and I don't want to do it.

Still need to implement functionality for background music.