Added some comments
Also made some minor changes & fixes

Fixed a bug with the background music

There was a delay between the intro part of the music and the outro part of the music. This was because Invoke() is timescale dependent and audio clips playing isn't.
Forgot to revert the timescale used for testing

Fixed a bug with the interaction prompt and NPCs