Last bugfixes and polish
Invisible wall added between the chair and the wall to prevent the player from getting stuck between them.

PauseMenu to main menu bug fixed. The input action map was still for controlling the character not the comic.

Keyboard buttons now centered in text.

Setting the sprites for the controls now checks  if a controller is plugged in when a device is added or removed not only at the start.

Xray now reduces alpha even further depending on how many objects are obscuring the player.